+++ src/routes/Routes.tsx
@@ -3,6 +3,7 @@ import { PATHS } from "./paths";
 import { Acknowledgments } from "../pages/Acknowledgments";
 import { MasteryTracker } from "../pages/MasteryTracker";
 import { WeeklyTracker } from "../pages/WeeklyTracker";
+import { Settings } from "../pages/Settings";
 
 export const ROUTES: RouteObject[] = [
     {
@@ -12,6 +13,10 @@ export const ROUTES: RouteObject[] = [
     {
         path: PATHS.Weekly,
         element: <WeeklyTracker />,
+    },
+    {
+        path: PATHS.Settings,
+        element: <Settings />,
     },
     {
         path: PATHS.Acknowledgments,